requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.4.0
celery==5.1.0
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
distlib==0.3.1
Django==3.1.2
django-environ==0.4.5
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
factory-boy==3.1.0
Faker==4.14.0
filelock==3.0.12
identify==1.5.6
idna==2.10
isort==5.6.4
kombu==5.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodeenv==1.5.0
oauthlib==3.1.0
phonenumbers==8.12.12
Pillow==8.0.1
pre-commit==2.7.1
prompt-toolkit==3.0.18
pycodestyle==2.6.0
pycrypto==2.6.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.11
vine==5.0.0
virtualenv==20.4.6
wcwidth==0.2.5
wrapt==1.12.1
